Fixes the LOB creation exception on startup.
Hibernate tries to detect whether some JDBC features work or not in
the database driver. This one causes an exception that is too low
level for Spring to catch, and it bubbles up into the logs with a
truly gigantic stack trace. By adding an application YAML with this
very deeply buried parameter we can tell Hibernate not to test that
feature, which prevents the stack trace from appearing.