Support verbs that implicitly enquote the rest of the input.
Verbs that are "quoting" will make the interpreter treat the rest of the input as if it was surrounded by quotes. That makes it so the communication commands work as experienced MUDders would expect:

say some really interesting and insightful stuff

Produces:

You say 'some really interesting and insightful stuff'

Otherwise the entire sentence would be tokenized as a regular command and the game would have a lot of trouble trying to figure out what all those words were supposed to mean. Alternatively you'd have to remember to put quotes around anything you want to say to make it into a quoted string (which doesn't actually work yet).