Basic framework for in-game maps.
GameMap is the basic unit for sections of the world. Each GameMap is a rectangle of arbitrary size, and each "room" (in the MUD sense) is represented as a byte within the GameMap's byte array. The value of the byte will eventually correspond to a tile in a tileset that tells us what that room is.

The really clever part here is that each GameMap is a single row in the database and will be primarily accessed by ID, which can be very fast. Each map will have "portals" to other maps which is how we'll build out a full sized world. It's also how we'll be able to handle containers and inventories. If you get picked up by someone you could find yourself in a single tile map representing their inventory. If you're really big and pick up a house, the house's map could break its link to whatever map it was originally on and now link to your inventory. I think we'll find this is a very flexible system that could eventually allow for vehicles, carrying other creatures, and a reasonable way to handle the fact that MUD rooms aren't all the same "size" the way roguelike tiles are.

I preloaded the database with a very boring "default" map which demonstrates two things. First, a way to preload the world with hard-coded content. That first map could be a MUD school newbie area or something like that. It's also a good testing ground for the next step, which is to add actors (aka. MOBs or NPCs and players).