Very quick and dirty implementation of Actors
The SAY command now actually passes your message to others who are logged on at the same time!

Actors are the in-game representation of items and creatures. Until this point we've been able to log in but we effectively haven't had any "physical" representation within the game world. Recently I added GameMap which provides a place, and here are Actors that provide things. Now when you enter the game an Actor is created for you, and when you disconnect your Actor is deleted. When you talk the game can figure out if there are other Actors in the same location as you and pass your messages to them.

This code still needs tests before it can be merged, is very simple and fragile in some places. However, it demonstrates the concept and provides a framework that can be expanded on. For instance, we can allow Actors to move around to different spots on the GameMap and to see each other with LOOK.