Extract comm service out of SayCommand
We're going to need to send messages to Actors all over the place, and we're going to need many variations on how it works (e.g. send to everyone in the room except me, send to everyone in an n room radius...) so it makes sense to start a service where all those utilities can be centralized. For now we just need the one way of doing it, so let's not over-think things.