Extract player connection data out into a component object
On one hand this refactor means more typing because now you have to type `getConnection()` a lot and a little more complexity if you're looking at the database directly, but it also makes it easier to distinguish an NPC or item from a player. If `getConnection()` is null, you know it's not a player!

The inspiration for this kind of design is actually Unity3D where there is a central `GameObject` that is basically just a shell with fields for dozens of optional components. I envision `Actor` being essentially that. In the future there will be a module for NPC data like AI/behavior settings, and a module for item data as well as a few others. The presence or absence of those modules will tell you what kind of object it is and what capabilities it has.