Fixed broken message dispatching
When trying to send messages from one application instance to the other over the broker, they were instead getting dropped. I had borrowed some code from EmergentMUD that set a session ID in the header of every message, which works great when there is a single application server. Removing the headers made the messages successfully travel between application servers, but when two tabs were logged in on the same account both tabs would get messages not intended for them; Spring can send messages to a /user/ but not so easily to a specific /session/.

The solution I eventually landed on was to stop setting the headers on the messages, and to implement UniqueHandshakeHandler which produces Principals with random UUIDs as usernames. Now that the users are all unique, Spring can send messages to individual characters without bleeding over into other tabs!