Base room descriptions
These will probably go away eventually in favor of completely generated room descriptions, but for now it does a lot for the game to have at least a little flavor text in each room. We'll have to experiment with whether a blend of some hand-written and some generated text is better or if we can do it all with generated text without making the world seem too bland.