Triggered NPE trying to make items LOOK
The LOOK command didn't check for actor.getCreatureInfo() being null before dereferencing it. GET and DROP make the Actor that was picked up or dropped execute a LOOK. Most of the time it's wasted effort since it's just an item but when it's a player it makes everything look right. When it was not a player, there was no CreatureInfo attached to the Actor and it threw the NPE. I somewhat over-aggressively added tests and null checks to prevent this from happening.