Enum indexes
The Apache EnumUtils used enum ordinals, which can shift unexpectedly when you add or subtract values from an enum. This change adds a PersistentEnum interface that offers a getIndex() method. The index can be set any way, but the pattern I have in mind is that you specify it in the enum's constructor. That way it will be very obvious what the value is and it will never change unless you change it yourself. That should help to avoid accidentally corrupting the database by adding new enums. The downside is that the converter can only convert PersistentEnums now, so you have to be deliberate when building your model classes.